Add files via upload
The idea is to use a csv file with word in any language and their english translations. Then, if the user knows the word and clicks the check or 'right' button that word won't come again. On the other hand, if the user doesn't know the word they can click the X or 'wrong' button and that word will stay on the desk. Once, all word from the file has been learned the program will notify with a pop-up box.