Add files via upload
This program uses Google Sheets and Tequila/Kiwi API to obtain flight information from the places listed in the spreadsheets. Also, could use the customer_acquisition to have user enter their information and be added to the user's spreadsheet. The program is able to email flight deals to the user's on the list, or use it to send SMS to a personal phone number via twilio.