Add files via upload
This program uses Google Sheets and Tequila/Kiwi APIs, to obtain flight deals according to the price point from the spreadsheets. If the IATA (airport) Code is not in the sheet it can also get it and update the spreadsheet but it'll need a reference price from the sheet as well. Also, the acquisition code gets a person's email and records it on the spreadsheet. Then, the program can send an SMS to a personal phone number with the information of each of the flights found, and it can send the emails to the users in the spreadsheet with a link to the booking website.